Bug 1796353 - Move URLStringUtils to ktx package.
This is needed to avoid circular dependencies for some needed changes.
